apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
requests==2.32.3
PyJWT==2.9.0
tree-sitter==0.23.2
tree-sitter-languages==1.10.2
